# Talent-Hubzip/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jobera.az Job Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job vacancies
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: employmentType
          in: query
          schema:
            type: string
        - name: salaryMin
          in: query
          schema:
            type: integer
        - name: salaryMax
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job vacancy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job vacancy details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job vacancy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job vacancy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── APPLICATIONS ───────────────────────────────────────────────────────────
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent activity feed for the current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ── AUTH ──────────────────────────────────────────────────────────────────
...
    Company:
      type: object
      required: [id, userId, name, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        employeeCount:
          type: ["string", "null"]
        isVerified:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
...
    Job:
      type: object
      required: [id, companyId, title, category, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        company:
          $ref: "#/components/schemas/Company"
        title:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        employmentType:
          type: string
          enum: [full_time, part_time, remote, hybrid, internship]
          default: full_time
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        currency:
          type: string
          default: AZN
        isActive:
          type: boolean
          default: true
        viewCount:
          type: integer
          default: 0
        applicationCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
    JobInput:
      type: object
      required: [title, category]
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        requirements:
          type: string
        city:
          type: string
        address:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, remote, hybrid, internship]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        currency:
          type: string
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        requirements:
          type: string
        city:
          type: string
        address:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, remote, hybrid, internship]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        currency:
          type: string
        isActive:
          type: boolean
    JobListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ActivityItem:
      type: object
      required: [id, type, title, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [application, contact_request, rating, job_posted, subscription]
        title:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time